backend/requirements.txt
face-recognition==1.3.0
numpy==1.24.2
Pillow==9.4.0
dlib==19.24.2
fastapi[all]
pymongo[srv]
python-dotenv
